Broke out the key.
To make sure I can't accidentally commit the key, I put it in `key.js`,
and made that an ignored file.

The readme has the info for recreating the file.

:shrug:

Yes, it is exploratory, but I know I'll commit a key. It's a superpower.
And, yes, I know we have gitleaks running. But still. :)